adapt-bio/requirements.txt
fastapi
pydantic
uvicorn
python-multipart
requests
farm-haystack
mlflow
boto3